+++ ark/schema/ast.ts
@@ -12,15 +12,9 @@ export type DateLiteral<source extends string = string> =
 	| `d"${source}"`
 	| `d'${source}'`
 
-export type Constraints = {
-	divisor?: { [k: number]: 1 }
-	min?: { [k: number | string]: 0 | 1 }
-	max?: { [k: number | string]: 0 | 1 }
-	regex?: { [k: string]: 1 }
-	length?: { [k: number]: 1 }
-	predicate?: 1
-	literal?: string | number
-}
+export type ConstraintSet = Record<PropertyKey, 1>
+
+export type Constraints = Record<string, ConstraintSet>
 
 export declare const constrained: unique symbol
 
@@ -37,40 +31,46 @@ export type LimitLiteral = number | DateLiteral
 export type normalizeLimit<limit> =
 	limit extends DateLiteral<infer source> ? source
 	: limit extends number | string ? limit
-	: string
+	: never
 
-export type AtLeast<rule extends number> = {
-	min: { [k in rule]: 0 | 1 }
+type constraint<rule> = { [k in rule & PropertyKey]: 1 }
+
+export type AtLeast<rule> = {
+	atLeast: constraint<rule>
 }
 
-export type AtMost<rule extends number> = {
-	max: { [k in rule]: 0 | 1 }
+export type AtMost<rule> = {
+	atMost: constraint<rule>
 }
 
-export type MoreThan<rule extends number> = {
-	min: { [k in rule]: 0 }
+export type MoreThan<rule> = {
+	moreThan: constraint<rule>
 }
 
-export type LessThan<rule extends number> = {
-	max: { [k in rule]: 0 }
+export type LessThan<rule> = {
+	lessThan: constraint<rule>
 }
 
-export type Literal<rule extends string | number> = { literal: rule }
-
-export type DivisibleBy<rule extends number> = {
-	divisor: { [k in rule]: 1 }
+export type Literal<rule> = {
+	literal: constraint<rule>
 }
 
-export type Length<rule extends number> = {
-	length: { [k in rule]: 1 }
+export type DivisibleBy<rule> = {
+	divisibleBy: constraint<rule>
 }
 
-export type Matching<rule extends string> = {
-	regex: { [k in rule]: 1 }
+export type Length<rule> = {
+	length: constraint<rule>
 }
+
+export type Matching<rule> = {
+	matching: constraint<rule>
+}
+
+export type anonymous = "?"
 
 export type Narrowed = {
-	predicate: 1
+	predicate: { [k in anonymous]: 1 }
 }
 
 export type primitiveConstraintKindOf<In> = Extract<
@@ -79,15 +79,15 @@ export type primitiveConstraintKindOf<In> = Extract<
 >
 
 export namespace number {
-	export type atLeast<rule extends number> = of<number, AtLeast<rule>>
+	export type atLeast<rule> = of<number, AtLeast<rule>>
 
-	export type moreThan<rule extends number> = of<number, MoreThan<rule>>
+	export type moreThan<rule> = of<number, MoreThan<rule>>
 
-	export type atMost<rule extends number> = of<number, AtMost<rule>>
+	export type atMost<rule> = of<number, AtMost<rule>>
 
-	export type lessThan<rule extends number> = of<number, LessThan<rule>>
+	export type lessThan<rule> = of<number, LessThan<rule>>
 
-	export type divisibleBy<rule extends number> = of<number, DivisibleBy<rule>>
+	export type divisibleBy<rule> = of<number, DivisibleBy<rule>>
 
 	export type narrowed = of<number, Narrowed>
 
@@ -100,62 +100,50 @@ export namespace number {
 		normalizePrimitiveConstraintRoot<schema> extends infer rule ?
 			kind extends "min" ?
 				schema extends { exclusive: true } ?
-					moreThan<rule & number>
-				:	atLeast<rule & number>
+					moreThan<rule>
+				:	atLeast<rule>
 			: kind extends "max" ?
 				schema extends { exclusive: true } ?
-					lessThan<rule & number>
-				:	atMost<rule & number>
-			: kind extends "divisor" ? divisibleBy<rule & number>
+					lessThan<rule>
+				:	atMost<rule>
+			: kind extends "divisor" ? divisibleBy<rule>
 			: narrowed
 		:	never
 }
 
-export type AtLeastLength<rule extends number> = {
-	min: { [k in rule]: 0 | 1 }
+export type AtLeastLength<rule> = {
+	atLeastLength: constraint<rule>
 }
 
-export type AtMostLength<rule extends number> = {
-	max: { [k in rule]: 0 | 1 }
+export type AtMostLength<rule> = {
+	atMostLength: constraint<rule>
 }
 
-export type MoreThanLength<rule extends number> = {
-	min: { [k in rule]: 0 }
+export type MoreThanLength<rule> = {
+	moreThanLength: constraint<rule>
 }
 
-export type LessThanLength<rule extends number> = {
-	max: { [k in rule]: 0 }
+export type LessThanLength<rule> = {
+	lessThanLength: constraint<rule>
 }
 
-export type ExactlyLength<rule extends number> = {
-	min: { [k in rule]: 1 }
-	max: { [k in rule]: 1 }
+export type ExactlyLength<rule> = {
+	atLeastLength: constraint<rule>
+	atMostLength: constraint<rule>
 }
 
 export namespace string {
-	export type atLeastLength<rule extends number> = of<
-		string,
-		AtLeastLength<rule>
-	>
+	export type atLeastLength<rule> = of<string, AtLeastLength<rule>>
 
-	export type moreThanLength<rule extends number> = of<
-		string,
-		MoreThanLength<rule>
-	>
+	export type moreThanLength<rule> = of<string, MoreThanLength<rule>>
 
-	export type atMostLength<rule extends number> = of<string, AtMostLength<rule>>
+	export type atMostLength<rule> = of<string, AtMostLength<rule>>
 
-	export type lessThanLength<rule extends number> = of<
-		string,
-		LessThanLength<rule>
-	>
+	export type lessThanLength<rule> = of<string, LessThanLength<rule>>
 
-	export type exactlyLength<rule extends number> = of<
-		string,
-		ExactlyLength<rule>
-	>
+	export type exactlyLength<rule> = of<string, ExactlyLength<rule>>
 
-	export type matching<rule extends string> = of<string, Matching<rule>>
+	export type matching<rule> = of<string, Matching<rule>>
 
 	export type narrowed = of<string, Narrowed>
 
@@ -168,52 +156,46 @@ export namespace string {
 		normalizePrimitiveConstraintRoot<schema> extends infer rule ?
 			kind extends "minLength" ?
 				schema extends { exclusive: true } ?
-					moreThanLength<rule & number>
-				:	atLeastLength<rule & number>
+					moreThanLength<rule>
+				:	atLeastLength<rule>
 			: kind extends "maxLength" ?
 				schema extends { exclusive: true } ?
-					lessThanLength<rule & number>
-				:	atMostLength<rule & number>
-			: kind extends "regex" ? matching<rule & string>
-			: kind extends "exactLength" ? exactlyLength<rule & number>
+					lessThanLength<rule>
+				:	atMostLength<rule>
+			: kind extends "pattern" ? matching<rule & string>
+			: kind extends "exactLength" ? exactlyLength<rule>
 			: narrowed
 		:	never
 }
 
-export type AtOrAfter<rule extends number | string> = {
-	min: { [k in rule]: 0 | 1 }
+export type AtOrAfter<rule> = {
+	atOrAfter: constraint<rule>
 }
 
-export type AtOrBefore<rule extends number | string> = {
-	max: { [k in rule]: 0 | 1 }
+export type AtOrBefore<rule> = {
+	atOrBefore: constraint<rule>
 }
 
-export type After<rule extends number | string> = {
-	min: { [k in rule]: 0 }
+export type After<rule> = {
+	after: constraint<rule>
 }
 
-export type Before<rule extends number | string> = {
-	max: { [k in rule]: 0 }
+export type Before<rule> = {
+	before: constraint<rule>
 }
 
 export namespace Date {
-	export type atOrAfter<rule extends string | number> = of<
-		Date,
-		AtOrAfter<rule>
-	>
+	export type atOrAfter<rule> = of<Date, AtOrAfter<rule>>
 
-	export type after<rule extends string | number> = of<Date, After<rule>>
+	export type after<rule> = of<Date, After<rule>>
 
-	export type atOrBefore<rule extends string | number> = of<
-		Date,
-		AtOrBefore<rule>
-	>
+	export type atOrBefore<rule> = of<Date, AtOrBefore<rule>>
 
-	export type before<rule extends string | number> = of<Date, Before<rule>>
+	export type before<rule> = of<Date, Before<rule>>
 
 	export type narrowed = of<Date, Narrowed>
 
-	export type literal<rule extends string | number> = of<Date, Literal<rule>>
+	export type literal<rule> = of<Date, Literal<rule>>
 
 	export type is<constraints extends Constraints> = of<Date, constraints>
 
@@ -265,34 +247,34 @@ type _constrain<
 export type normalizePrimitiveConstraintRoot<
 	schema extends NodeSchema<PrimitiveConstraintKind>
 > =
-	"rule" extends keyof schema ? conform<schema["rule"], string | number>
-	:	conform<schema, string | number>
+	"rule" extends keyof schema ? conform<schema["rule"], PropertyKey>
+	:	conform<schema, PropertyKey>
 
 export type schemaToConstraint<
 	kind extends PrimitiveConstraintKind,
 	schema extends NodeSchema<kind>
 > =
 	normalizePrimitiveConstraintRoot<schema> extends infer rule ?
-		kind extends "regex" ? Matching<rule & string>
-		: kind extends "divisor" ? DivisibleBy<rule & number>
-		: kind extends "exactLength" ? Length<rule & number>
+		kind extends "pattern" ? Matching<rule>
+		: kind extends "divisor" ? DivisibleBy<rule>
+		: kind extends "exactLength" ? Length<rule>
 		: kind extends "min" ?
 			schema extends { exclusive: true } ?
-				MoreThan<rule & number>
-			:	AtLeast<rule & number>
+				MoreThan<rule>
+			:	AtLeast<rule>
 		: kind extends "max" ?
 			schema extends { exclusive: true } ?
-				LessThan<rule & number>
-			:	AtMost<rule & number>
+				LessThan<rule>
+			:	AtMost<rule>
 		: kind extends "minLength" ?
 			schema extends { exclusive: true } ?
-				MoreThanLength<rule & number>
-			:	AtLeastLength<rule & number>
+				MoreThanLength<rule>
+			:	AtLeastLength<rule>
 		: kind extends "maxLength" ?
 			schema extends { exclusive: true } ?
-				LessThanLength<rule & number>
-			:	AtMostLength<rule & number>
-		: kind extends "exactLength" ? ExactlyLength<rule & number>
+				LessThanLength<rule>
+			:	AtMostLength<rule>
+		: kind extends "exactLength" ? ExactlyLength<rule>
 		: kind extends "after" ?
 			schema extends { exclusive: true } ?
 				After<normalizeLimit<rule>>