+++ ark/schema/roots/root.ts
@@ -6,7 +6,7 @@ import type {
 	InclusiveDateRangeSchema,
 	InclusiveNumericRangeSchema,
 	LimitSchemaValue,
-	RegexSchema,
+	PatternSchema,
 	UnknownRangeSchema
 } from "@arktype/schema"
 import {
@@ -86,7 +86,7 @@ export abstract class BaseRoot<
 	implements internalImplementationOf<Root, TypeOnlyRootKey | "intersect">
 {
 	readonly branches: readonly Node<UnionChildKind>[] =
-		this.hasKind("union") ? this.inner.branches : [this as never];
+		this.hasKind("union") ? this.inner.branches : [this as never]
 
 	readonly [arkKind] = "root"
 
@@ -282,8 +282,8 @@ export abstract class BaseRoot<
 		return this.constrain("divisor", schema)
 	}
 
-	matching(schema: RegexSchema): BaseRoot {
-		return this.constrain("regex", schema)
+	matching(schema: PatternSchema): BaseRoot {
+		return this.constrain("pattern", schema)
 	}
 
 	atLeast(schema: InclusiveNumericRangeSchema): BaseRoot {